Restructure the whole directory structure based other contributors feedback.
