+++ framework/internals/async/async.go
@@ -11,7 +11,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"bean/internals/global"
+	"bean/framework/internals/global"
 
 	"github.com/getsentry/sentry-go"
 	sentryecho "github.com/getsentry/sentry-go/echo"
