+++ framework/internals/global/constants.go
