+++ framework/internals/stacktrace/stack.go
