+++ framework/internals/struct/structure.go
