+++ repositories/mytest_repository.go
@@ -6,7 +6,7 @@
 package repositories
 
 import (
-	"bean/internals/global"
+	"bean/framework/internals/global"
 
 	"github.com/labstack/echo/v4"
 )
